Folder restructure and fixes packages
